refactor: structure repo, detach backendapi to folders: api, utils, model, add folder docs to contain API Documentation, add folder tests for pytest
